add state wrapper to test utils
