+++ src/tests/NoteView.test.js
@@ -1,39 +1,24 @@
 import * as ReactRouter from 'react-router';
 import NoteView from "../components/NoteView/NoteView";
 import {render, screen, getByText} from "@testing-library/react";
-import {wrappedRender} from "./testUtils";
+import {wrappedRender, wrapWithAppState} from "./testUtils";
+import React, { Children, useState } from 'react';
+
+
+
+
 
 it("displays data correct", ()=>{
     //arrange
-    const setEditedNote=jest.fn();
-    const editedNote=jest.fn();
-    const setFavouriteShortened=jest.fn();
-    const notes=[{
-        "id": 1,
-        "title": "test1",
-        "content": "test1 content here",
-        "isFavourite": false,
-        "description": "description test"
-      },
-      {
-        "id": 2,
-        "title": "example",
-        "content": "exaple example",
-        "isFavourite": false,
-        "description": "example note"
-    }];
+
+    const StateWrappedNoteView=wrapWithAppState(NoteView);
 
     //act
-
     jest.spyOn(ReactRouter, 'useParams').mockReturnValue({"*": 1});
-
     wrappedRender(
-        <NoteView 
-         setEditedNote={setEditedNote}
-         editedNote={editedNote}
-         setFavourite={setFavouriteShortened}
-         notes={notes}
-        />
+        <StateWrappedNoteView />
     )
+
     //assert
+    expect(screen.getByRole("textbox")).toHaveTextContent("test1 content here");
 })
